avance de apis

+++ apisRestUdemy/proyecto02/api-rest-cuentas/src/main/java/com/api/hateoas/service/CuentaService.java
@@ -1,6 +1,7 @@
 package com.api.hateoas.service;
 
 
+import com.api.hateoas.exception.CuentaNotFoundException;
 import com.api.hateoas.model.Cuenta;
 import com.api.hateoas.repository.CuentaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,28 +17,35 @@ public class CuentaService {
     @Autowired
     private CuentaRepository cuentaRepository;
 
-    public List<Cuenta> listAll(){
+    public List<Cuenta> listAll() {
         return cuentaRepository.findAll();
     }
 
-    public Cuenta get(Integer id){
+    public Cuenta get(Integer id) {
         return cuentaRepository.findById(id).get();
     }
 
-    public Cuenta save(Cuenta cuenta){
+    public Cuenta save(Cuenta cuenta) {
         return cuentaRepository.save(cuenta);
     }
 
-    public void delete(Integer id){
+    public void delete(Integer id) throws CuentaNotFoundException {
+
+        if (!cuentaRepository.existsById(id)) {
+
+            throw new CuentaNotFoundException("Cuenta no encontrada con el ID:" + id);
+        }
         cuentaRepository.deleteById(id);
+
+
     }
 
-    public Cuenta depositar(float monto,Integer id){
+    public Cuenta depositar(float monto, Integer id) {
         cuentaRepository.actualizarMonto(monto, id);
         return cuentaRepository.findById(id).get();
     }
 
-    public Cuenta retitar(float monto,Integer id){
+    public Cuenta retitar(float monto, Integer id) {
         cuentaRepository.actualizarMonto(-monto, id);
         return cuentaRepository.findById(id).get();
     }
